requirements.txt
tensorflow 
streamlit 
matplotlib 
pillow 
numpy